refactor: deconstruct monolithic report service into a package
